requirements.txt
streamlit==1.28.2
opencv-python==4.7.0.72
tensorflow==2.16.1
numpy
Pillow
